Fix issue in which trying to mouse select a text input caused the element to drag.
The issue with how the draggable list is used is that it encompasses all elements within the container. This means that if we drag anywhere, it triggers.

This is not an ideal condition when dealing with text inputs. It is often that case that you'll drag over the text to highlight it to remove. Specifically because their is a default name given that likely all users are removing.

To fix this, remove the draggable attribute when the input is focused and add it back once the attribute is blurred.