fix: avoid invoking `Token` during render hooks
This avoids a case where the grid has not fully initialized, and so
attempting to create tokens throws an error when the token attempts to
read grid properties.

Or something like that, I'm not confident that this is the exact cause--
either way, it appears to be a difficult-to-reproduce race condition,
and this change sidesteps it by avoiding `Token()` (which was
ultimately not necessary anyway).